[arc] Move program termination analysis a utility in the ARC folder (for now).
I eventually am probably going to integrate SILCodeMotion into ARCLoopOpts at
which point I will probably need to refactor this into SILPasses/Utils.
