+++ include/swift/SILAnalysis/Analysis.def
@@ -35,7 +35,6 @@ ANALYSIS(Loop)
 ANALYSIS(LoopRegion)
 ANALYSIS(PostDominance)
 ANALYSIS(PostOrder)
-ANALYSIS(ProgramTermination)
 ANALYSIS(RCIdentity)
 ANALYSIS(SideEffect)
