+++ lib/SILPasses/Scalar/SILCodeMotion.cpp
@@ -24,7 +24,6 @@
 #include "swift/SILAnalysis/AliasAnalysis.h"
 #include "swift/SILAnalysis/PostOrderAnalysis.h"
 #include "swift/SILAnalysis/RCIdentityAnalysis.h"
-#include "swift/SILAnalysis/ProgramTerminationAnalysis.h"
 #include "swift/SILPasses/Transforms.h"
 #include "swift/SILPasses/Utils/Local.h"
 #include "llvm/ADT/Optional.h"
@@ -808,8 +807,7 @@ static bool tryToSinkRefCountAcrossSelectEnum(CondBranchInst *CondBr,
 static bool tryToSinkRefCountInst(SILBasicBlock::iterator T,
                                   SILBasicBlock::iterator I,
                                   bool CanSinkToSuccessors, AliasAnalysis *AA,
-                                  RCIdentityFunctionInfo *RCIA,
-                                  ProgramTerminationFunctionInfo *PTFI) {
+                                  RCIdentityFunctionInfo *RCIA) {
   // The following methods should only be attempted if we can sink to our
   // successor.
   if (CanSinkToSuccessors) {
@@ -857,7 +855,7 @@ static bool tryToSinkRefCountInst(SILBasicBlock::iterator T,
   for (auto &Succ : T->getParent()->getSuccessors()) {
     SILBasicBlock *SuccBB = Succ.getBB();
 
-    if (PTFI->isProgramTerminatingBlock(SuccBB))
+    if (isARCInertTrapBB(SuccBB))
       continue;
 
     Builder.setInsertionPoint(&*SuccBB->begin());
@@ -986,8 +984,7 @@ static bool hoistDecrementsToPredecessors(SILBasicBlock *BB, AliasAnalysis *AA,
 /// Try sink a retain as far as possible.  This is either to sucessor BBs,
 /// or as far down the current BB as possible
 static bool sinkRefCountIncrement(SILBasicBlock *BB, AliasAnalysis *AA,
-                                  RCIdentityFunctionInfo *RCIA,
-                                  ProgramTerminationFunctionInfo *PTFI) {
+                                  RCIdentityFunctionInfo *RCIA) {
 
   // Make sure that each one of our successors only has one predecessor,
   // us.
@@ -1021,12 +1018,12 @@ static bool sinkRefCountIncrement(SILBasicBlock *BB, AliasAnalysis *AA,
     //   2. If there are such decrements, move the retain right before that
     //      decrement.
     Changed |= tryToSinkRefCountInst(S->getIterator(), Inst->getIterator(),
-                                     CanSinkToSuccessor, AA, RCIA, PTFI);
+                                     CanSinkToSuccessor, AA, RCIA);
   }
 
   // Handle the first instruction in the BB.
-  Changed |= tryToSinkRefCountInst(S->getIterator(), SI, CanSinkToSuccessor, AA,
-                                   RCIA, PTFI);
+  Changed |=
+      tryToSinkRefCountInst(S->getIterator(), SI, CanSinkToSuccessor, AA, RCIA);
   return Changed;
 }
 
@@ -1672,7 +1669,6 @@ sinkIncrementsOutOfSwitchRegions(AliasAnalysis *AA,
 static bool processFunction(SILFunction *F, AliasAnalysis *AA,
                             PostOrderFunctionInfo *PO,
                             RCIdentityFunctionInfo *RCIA,
-                            ProgramTerminationFunctionInfo *PTFI,
                             bool HoistReleases) {
 
   bool Changed = false;
@@ -1724,7 +1720,7 @@ static bool processFunction(SILFunction *F, AliasAnalysis *AA,
     Changed |= State.process();
 
     // Finally we try to sink retain instructions from this BB to the next BB.
-    Changed |= sinkRefCountIncrement(State.getBB(), AA, RCIA, PTFI);
+    Changed |= sinkRefCountIncrement(State.getBB(), AA, RCIA);
 
     // And hoist decrements to predecessors. This is beneficial if we can then
     // match them up with an increment in some of the predecessors.
@@ -1749,12 +1745,11 @@ public:
     auto *AA = getAnalysis<AliasAnalysis>();
     auto *PO = getAnalysis<PostOrderAnalysis>()->get(F);
     auto *RCIA = getAnalysis<RCIdentityAnalysis>()->get(getFunction());
-    auto *PTFI = getAnalysis<ProgramTerminationAnalysis>()->get(getFunction());
 
     DEBUG(llvm::dbgs() << "***** CodeMotion on function: " << F->getName() <<
           " *****\n");
 
-    if (processFunction(F, AA, PO, RCIA, PTFI, HoistReleases))
+    if (processFunction(F, AA, PO, RCIA, HoistReleases))
       invalidateAnalysis(SILAnalysis::InvalidationKind::Instructions);
   }
